Moves test classes to separate files
