+++ cerberus/tests/test_legacy.py
@@ -1,0 +1,1 @@
+pass
